feat: Support indentation of yara code through smie
`smie' is an emacs core package providing a generic navigation and
indentation engine. It makes emacs' auto-indentation works, as well as
the sexp-based operations, such as forward-sexp and backward-sexp.

`smie' requires a precedence table for automating indentation, which
is suggested to be created from a concise BNF of the target language
(yara, in this commit). Note that BNF is used for indentation only,
hence no need to be exactly the same as the real yara syntax, and
that is in fact what this commit does -- only modeling the toplevel
statements, such as import/include/rule statements, and the section
statements inner the rule block -- for the sake of efficiency.